refactor: create vite-plugin-symlink
