+++ apps/outreach/package.json
@@ -21,13 +21,14 @@
   },
   "devDependencies": {
     "@astrojs/check": "^0.7.0",
-    "@astrojs/sitemap": "^3.1.5",
-    "@astrojs/starlight": "^0.24.1",
+    "@astrojs/sitemap": "^3.1.6",
+    "@astrojs/starlight": "^0.24.4",
     "@astrojs/tailwind": "^5.1.0",
     "@opendatacapture/tailwindcss": "workspace:*",
+    "@opendatacapture/vite-plugin-symlink": "workspace:*",
     "@tailwindcss/typography": "^0.5.13",
     "@types/lodash-es": "^4.17.12",
-    "astro": "^4.10.1",
+    "astro": "^4.11.1",
     "sharp": "^0.33.4",
     "tailwindcss": "^3.4.4"
   }
